Modified program structure to allow configurable number of threads at runtime.
Program now takes command line arguments of the form -t<int> (for maximum number of worker threads to use) and -d[Detailed/Sample/Basic] for level of logging.

Also fixed a bug where the ballast wasn't being cleared when transitioning back to the main menu. And fixed a race condition in the data parallel version where creeps might not get added to the main list in certain conditions. Also hopefully fixed a hang on transitioning back to the main menu from the task based version in certain situations.

Signed-off-by: Andrew James <andrewianjames@gmail.com>